containers: Add `onActivityExpired` so users can configured how to act on activity expiration
Activity expiration now is measured since the last time a HTTP request
has arrived to the container class. We can't reliably measure when a
response gets closed to renew the activity right now, so let's rely for
now on the user to decide if they want to renew activity by optionally
contacting with the container.